Added a REcepieIndex component and got the SustainBadge to stick on top of the Card, somehow lost the scrollView
