+++ .expo/packager-info 3.json
@@ -1,9 +1,9 @@
 {
   "devToolsPort": 19002,
-  "expoServerPort": null,
-  "packagerPort": null,
-  "packagerPid": null,
+  "expoServerPort": 19000,
+  "packagerPort": 19001,
+  "packagerPid": 73523,
   "expoServerNgrokUrl": "https://sd-vdy.anonymous.final-project-front-end.exp.direct",
   "packagerNgrokUrl": "https://packager.sd-vdy.anonymous.final-project-front-end.exp.direct",
-  "ngrokPid": 76596
+  "ngrokPid": 73533
 }
