+++ Components/RecepieCard.js
@@ -1,17 +1,13 @@
 import React from 'react';
-import { ImageBackground, Text, View } from 'react-native'
+import { ImageBackground, Text } from 'react-native'
 import { Card, Title, Paragraph } from 'react-native-paper';
 import TimeComponent from './TimeComponent'
-import SustainBadge from './SustainBadge'
 import styled from 'styled-components/native'
 
 
 const RecepieCard = () => (
     <StyledCard>
         <Card.Content>
-            <BadgeView>
-                <SustainBadge />
-            </BadgeView>
             <StyledImageBackground source={require('../assets/burger-and-vegetables-placed-on-brown-wood-surface-1565982.jpg')}>
                 <StyledTitle>"Here goes the title for the recepie {}"</StyledTitle>
                 <StyledText>"Card Subtitle{}"</StyledText>
@@ -40,18 +36,9 @@ box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
 z-index: 0;
 `;
 
-const BadgeView = styled.View`
-z-index: 2;
-position: absolute;
-left: 1px;
-top: 1px;
-`;
-
 const StyledImageBackground = styled(ImageBackground)`
 width: 100%;
 height: 200px;
-
-
 border-radius: 30;
 z-index:1;
 position: absolute;
@@ -59,11 +46,9 @@ left: 8%;
 right: 0%;
 top: 60px;
 bottom: 0%;
-
 `;
 
 const StyledTitle = styled(Title)`
-
 color: #FFFFFF;
 position: absolute;
 left: 0%;
