+++ package.json
@@ -9,7 +9,9 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.2.5",
+    "braces": "^3.0.2",
     "expo": "^35.0.0",
+    "mem": "^6.0.1",
     "react": "16.8.3",
     "react-dom": "16.8.3",
     "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
